Fix #3: rename
