+++ rslaser/utils/string_tools.py
